+++ CHANGELOG.md
@@ -1,3 +1,27 @@
+## 0.111.0 (March 4, 2022)
+
+### Features
+
+- Implements optional scenarios for backstopJS.
+
+## 0.110.2 (March 3, 2022)
+
+### Features
+
+- Inherit stdio setting for `StyleguideTester`.
+
+## 0.110.1 (March 3, 2022)
+
+### Maintenance
+
+- Cleanup `StyleguideTester` worker configuration.
+
+## 0.110.0 (March 3, 2022)
+
+### Features
+
+- Implements `StyleguideTester` worker to enable Styleguide Snapshot testing.
+
 ## 0.105.2 (February 21, 2022)
 
 ### Maintenance